yield exprs [2/n] extract code from forward_statement
Summary:
THe challenge in properly typing generator code is that we've
packed too much logic into our `forward_statement` function.

My initial priority is to correctly handle type resolution
of `yield` expressions. This requires the ability to inspect
the return annotation inside of `forward_expression`, so I'm
moving that logic into a separate (private) function in the
`State` module.

Reviewed By: grievejia

Differential Revision: D32288480

fbshipit-source-id: eae207b6c5cfad93e4384e6b6799cfed2cb1f595