Do not store all parsed CST modules in memory for statistics
Summary:
We're seeing OOMs on a couple large projects in fbcode, resulting in no statistics gathered and empty overview in our coverage UI.

It looks like the OOM is happening when we iterate through all the paths we care about, parse the text and create a libcst module object. We try to store all of these modules in memory, before running stats gathering logic over these in sequence.

Instead, let's avoid storing the CST objects all at once and just gather stats on each individual one as we parse them, so we just store the final counts as we go.

Reviewed By: grievejia

Differential Revision: D32321860

fbshipit-source-id: aa324d8a8d4528d921587633559ea81267d2fd27