Do not log annotations for path on libcst recursion error
Summary:
Some cleanup to avoid passing path unnecessarily into the libcst visitor helpers.

Since the recursion error is thrown on the .visit() call to module which is happening regardless of which type of collector we're running, we can probably even avoid a bunch of different try/catch blocks and just do one wrapping block around the toplevel caller.

I also think it could make more sense to not report for this path at all rather than report empty numbers for it, which is a semantic change in this diff.

Reviewed By: grievejia

Differential Revision: D32339428

fbshipit-source-id: f3ed94b6627f49c8753b704723287f14fe84c51a