yield validation [6/n] handle generator returns
Summary:
According to PEP 380, generators have a return type in addition
to a send and yield type. This applies only to regular generators,
PEP 525 states that async generators do not support `yield from`
syntax and do not permit non-empty return statements (empty return
is fine, that's just a way to terminate the generator).

So, in the case of an explicit return of a non-async generator,
we need to translate return values into an appropriate Generator
type before calling `validate_return`.

To the best of my knowledge, this closes off all of the remaining
bugs in our generator type checking. Generators are a bit complicated
so there may still be some edge cases we miss (I very nearly wrote
this in a commit where we would have failed to throw an error when
returning None from a generator whose return type is non-None, which
goes to show that it's easy to hit edge cases!) but we're in much
better shape than we were.

Reviewed By: grievejia

Differential Revision: D32318431

fbshipit-source-id: 5a43bf4b98b40bf1e08b55c9c45fc85c8a68e5c1