yield validation [4/n] handle yield types correctly
Summary:
The goal of this commit is to prevent us from throwing
spurious errors when checking `yield` and `yield from`
expressions due to being too restrictive in our type
constraints.

The underlying problem is that the `Type.generator`
function was trying to do way too much. I've now split
it's functionality into three distinct functions:
- `Type.generator_expression` is equivalent to the old
  `Type.generator`: it produces the correct type for a
  generator expression (i.e. a comprehension; it's called
  a generator in the python grammar though). These generators
  are effectively just iterators, but they do have a `send`
  method (which enforces that only `None` is allowed).
- `Type.generator` takes three different optional type
  arguments for the yield type, send type, and return type.
  Anything not provided is defaulted to `type.Any`. The
  point here is to be able to restrict only the portions
  of the define return anntoation that we know based on a
  specific expression. Setting the other types to `Any`
  prevents us from incorrectly complaining, for example,
  if we `yield 7` in a `Generator[int, str, float]` because
  the type isn't exactly `Generator[int, None, None]`
- `Type.async_generator` works the same, but only permits
  us to set the `send_type` and `yield_type`; async generators
  have no return type.

As a result of this we correctly restrict generators
expressions to have `None` as the send and return types,
but are appropriately flexible when typing `yield` expressions.

**Where are we at this point?**

As of this commit, we no longer return any spurious errors in
generators except when they have an explicit return statement.
We're still missing a check of the send type in `yield from`.
I'll fix these remaining issues in upcoming commits.

Reviewed By: grievejia

Differential Revision: D32301237

fbshipit-source-id: eb52635267d536e11e46939a49bce9d3665a17ab