Preprocessing: qualify match subject
Summary:
# This diff

Adds qualification preprocessing of the match subject.

The context below is not so important for this particular diff, but as it will be form the bottom of this stack, putting it here for high-level details.

# Stack/Context

This stack implements PEP 634: structural pattern matching in Pyre (also known as the match statement).
https://www.python.org/dev/peps/pep-0634/

The overall approach is similar & consistent with how if, while etc are handled. Those are converted into a combination of control flow structures; and the condition put into an Assert statement in the appropriate block. This means that type-check doesn't not need additional special handling.

The translation of the pattern-match into an assert is handled by the MatchHelper module introduced here. This is done for code clarity/separation. We still keep it in the same file, since this is the only place it is used right now.  It also means that the interface of cfg module is not impacted.

Reviewed By: grievejia

Differential Revision: D32321743

fbshipit-source-id: b68c6478ccc9b8e3976d5d63962b45603c5bb167