Add DecoratorFactory to pyre_extensions (#525)
Summary:
Pull Request resolved: https://github.com/facebook/pyre-check/pull/525

The [Pyre documentation](https://www.internalfb.com/intern/staticdocs/pyre/docs/errors/#decorator-factories) shows us how to make a decorator factory using a protocol class.

But that means we're going to have a lot of protocol classes in the codebase all serving the same purpose.

Maybe it would be better to define a Pyre extension to get the job done?

Reviewed By: pradeep90

Differential Revision: D32150175

fbshipit-source-id: 0271b7eb98c0dc45b72015c4a7db767b195a2577