Adds key validation to taint.config files (#529)
Summary:
Adds validation of required and optional keys of taint.config files to
taintConfiguration parser in the server.

Previosuly, we didn't do such validation and ignored invalid keys in the
file and errored out with "not so pretty" error message when a required
key was absent.

Now we error out with a pretty error message and show a warning about
invalid keys while continuing through the parsing process.

Screenshot:
<img width="1440" alt="Screenshot 2021-11-05 at 11 30 47 PM" src="https://user-images.githubusercontent.com/8947010/140559648-62785475-221d-49da-b0aa-c517ebf2c1b6.png">

Pull Request resolved: https://github.com/facebook/pyre-check/pull/529

Test Plan:
- Compile the server from source
- add an (or some) invalid key(s) to your taint.config file
- run pysa with PYRE_BINARY pointing to the executable build from source
- see the invalid key message
- remove a (or some) required key(s) from your taint.config file
- run pysa with PYRE_BINARY pointing to the executable build from source
- see the pretty new error.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>
Fixes: https://github.com/MLH-Fellowship/pyre-check/issues/87

Reviewed By: grievejia

Differential Revision: D32213899

Pulled By: arthaud

fbshipit-source-id: d908218372ea37e015b43a666904d008b2564e63