Infer: handle directories in `--paths-to-modify`
Summary:
The old `pyre infer` could handle `--paths-to-modify` specifiying
a directory.

In order to deal with timeouts we were seeing in codemods, we modified
the logic to filter paths in the ocaml backend, so that we could do
far less analysis on big projects. But the way we did the checking
assumed that `paths_to_modify` would only contain actual file paths,
not directories.

This commit adds directory support to the ocaml-side qualifier filtering.

**Will ocaml and python work together? Some notes**

This new logic relies on absolute paths, so that it's feasible
to recursively look up parent directories using `Filename.dirname` until
we hit root. The output of `filename_lookup` is never relative, so this
is fine.

We also assume that `paths_to_modify` will always have absolute paths; we
would fail to detect relative paths. That's okay, the python-side logic
converts any relative paths into absolute paths at the entrypoint to `infer`
when we convert each raw path into `Path.cwd() / Path(path)`.

Finally, because `Filename.dirname` does not produce a trailing slash, we
assume that there are never trailing slashes in `paths_to_modify`. This
is also safe, I confirmed that `Path(path)` automatically strips trailing
slashes.

Reviewed By: grievejia

Differential Revision: D32361747

fbshipit-source-id: e7bd26e125a2b86d28a4c201d53e2212ca220384