Store the return type of a call in the call graph
Summary:
During the taint analysis, we need to resolve the return type of a call in order to add breadcrumbs.
This operation is quite costly and can be a bottleneck is some functions.
Let's compute the return type of a call only once and store it in the call graph.

Reviewed By: dkgi

Differential Revision: D32374934

fbshipit-source-id: df367b7ec5db13634d3752b7d54a81ae087c4f26