Refinement refactor [2/n] s/create/create_mutable/g
Summary:
The `Annotation` module has a general `create` that defaults
to creating a mutable anotation, and in fact the optional
`mutability` argument is never used.

It makes more sense to treat mutable and immutable annotations
in parallel by having two similarly named smart constructors,
and eliminate the unused optional argument.

The code does get a little more verbose because the more descriptive
name is longer, but I think it's worth it (and in the same spirit as
our "no abbreviations")

The diff looks big but it's just a search-and-replace + formatting.
No logic change whatsoever.

Reviewed By: grievejia

Differential Revision: D32401571

fbshipit-source-id: 5385d8ae25868154e4ec0d775c6286b24e95c1e8