Annotation refactor [4/n] add display_as_revealed_type
Summary:
As discussed in a previous commit, I'm trying to make the internal
details of `Annotation.t`s structure private, and eventually
turn it into a proper lattice which will help me significantly
simplify `RefinementUnit` code.

One of the places where we are (in my opinion inappropriately)
reaching into the guts of `Annotation.t` is the printing of
`RevealedType` errors. In this commit I make a `display_as_revealed_type`
function, whose logic is slightly different from `show` in ways
that do merit having a separate function, and use that for error
printing.

The results are identical.

Reviewed By: dkgi

Differential Revision: D32409423

fbshipit-source-id: f2a434d2a3ae1661791461297075e722933f349a