Annotation refactor [3/n] add transform_type
Summary:
The `Annotation.t` type contains multiple possible
slots for a `Type.t`. In several parts of the code,
we want to transform all of them the same way.

Instead of copy-pasting the same pattern match many
times, we can just abstract that operation into a function.

This is a slight readability win, but the real reason
I'm trying to do this is that I want to eventually hide
the internals of `Annotation.t` entirely (as in, other modules
should *never* be pattern matching on it, the precise structure
of the type shouldn't be public), and to do that
requires having reasonable abstractions.

Part of the context is that I am looking into extending
type refinement to be able to handle indexes, and the code is
currently messy because instead of smoothly propagating a
lattice from Type to Annotation to RefinementUnit to Resolution,
we are instead skipping over Annotation and coupling to its
implementation details. So, I'm trying to abstract the
implementation as well as make a proper lattice in this stack.

Reviewed By: dkgi

Differential Revision: D32403132

fbshipit-source-id: f6b1d25193d675a5d03ce484e2446613a2f52b9e