Annotation refactor [6/n] add join and meet
Summary:
Building on the previous commit, I move the `meet` and `join`
functions of the mutability-extended-type lattice in `Annotation.t`
into the module where it belongs.

The `RefinementUnit` code is still kind of hard to read but it's
significantly clearer than before, and more importantly at least we
can reason about whether the `Annotation.t` layer of the logic is correct
without worrying about all the map transformations in `RefinementUnit`.

Reviewed By: dkgi

Differential Revision: D32409421

fbshipit-source-id: f54294a23a5367b192ba2fe8bb48a6ad94626af3