RefinementUnit refactor [7] Fix copypasta bug in meet
Summary:
In the type system lattice, down means more information
and so `meet` should be be taking all of the attribute
refinements.

In fact, inside of the `fold2` logic that's what it's
doing: if `` `Left`` or `` `Right`` exists, we will take it
(unlike in `join` where we only take for `` `Both``).

This is a bug, caused partially because `meet` and `join`
are created by copy-pasting a lot of identical complicated
data processing logic, and then finding all of the lines
relevent to the math of lattices and editing them.

Reviewed By: dkgi

Differential Revision: D32419041

fbshipit-source-id: a4160e8d7ffd4e2b3160151d18d14a63274c8012