RefinementUnit refactor [1/n] Clean up less_or_equal
Summary:
The `less_or_equal` function was pretty hard to understand.

I simplified by cleanly separating it into two helper functions:
- are the bases less or equal?
- are the attributes less or equal? (uses a recursive call)

The outermost logic then becomes trivial and easy to read:
- compute whether the bases are less or equal
- if yes, then and only then compute whether the attributes are
  less or equal

The previous version of the function was mathematically
equivalent, but obfuscated enough that it was (a) hard to read
and (b) easy to miss that we were doing the recursive call even
when unecessary!

NOTE: This should be not only a readability win but a perf win.

Reviewed By: dkgi

Differential Revision: D32415207

fbshipit-source-id: cadd9360e90e53beb91c8f02d6d4df05f81d8da7