RefinementUnit refactor [10] renames
Summary:
The `meet` and `join` functions are actually just recursive
tree operations, but this is obfuscated to the point where it
took me on the order of a half hour to understand what they were
actually doing (and that's *after* moving the `Annotation.t`-level
lattice logic into a separate module).

In preparation for a refactor where I actually make them into
recursive functions, this commit renames things so that you
can look at the code and see what's actually going on:
- the `valid_*` functions were not actually doing any validation,
  I'm not sure where that name came from. They do two things: (1)
  they determine whether to short-circuit recursive processing
  of the rest of the tree and (2) they join or meet the base
  entry. Since it's hard to name a function that does two things,
  I just call this operation `process_root`.
- The output of the `process_root` function should therefore
  be called `should_recurse` or some similarly evocative name,
  rather than `valid` (which makes it sound like we're "validating"
  something).
- the `join/meet` and `join_refinement_unit/meet_refinement_unit`
  functions had unhelpful names: obviously we're joining or meeting
  things inside of `join` and `meet`, but the important thing is
  which *level* we're processing. I decided to use the same names
  in both functions since that makes it easier to see the common
  logic, I called the outer layer `combine_attributes` since
  it uses a fold to combine the `attributes` map, and the inner
  one `combine_one_attribute` since it handles the inner processing
  needed for `fold2`.

I'm going to rearrange everything in the next commit, this rename
commit is just because, if the names make sense, it will be much
easier to review the next commit.

Reviewed By: dkgi

Differential Revision: D32422080

fbshipit-source-id: 4c22ef05d673243d99688d222a6180c46887ad5d