RefinementUnit refactor [12] extract types_are_orderable
Summary:
We're wasting quite a bit of vertical space on doing exactly the same
two-way comparison of `left.annotation` and `right.annotation` in
the definition of `should_recurse`.

Also, the exact same operation comes up in the `typeCheck.ml`
refinement code.

Extracting this into a `GlobalResolution` function helps all around.

Reviewed By: dkgi

Differential Revision: D32422102

fbshipit-source-id: c46d32ef5aa87597d0657c34df9df75c74546751