RefinementUnit refactor [0/n] Move refine to GlobalResolution
Summary:
The `refine` function is just a wrapper around `Annotation.refine`, that
handles the boilerplate of setting up the type order function parameters.

It has nothing to do with `RefinementUnit.t`, and as a result doesn't
properly belong in the `RefinementUnit` module. Instead, we ought to keep
it in `GlobalResolution` (which is where we put all of the other type
order and type constraint functionality that extracts ordering info
from a `global_resolution` value).

Reviewed By: dkgi

Differential Revision: D32414541

fbshipit-source-id: 904431dd394a843d86740551336db7d8555c9507