Annotation refactor [7/n] move refine logic
Summary:
The function `RefinementUnit.refine` bothers me for a couple
of reasons
- first, the name and module are misleading - it has nothing to
  do with `RefinementUnit.t`, it's an operation on `Annotation.t`
- second, it's the only remaining complex logic that reaches
  into the guts of `Annotation.t`, breaking encapsulation.

So, (much as with the Annotation.t lattice functions), I'm
pulling the core logic into Annotation.t, and using higher-order
functions to inject the necessary type lattice functionality.

I want to move `RefinementUnit.refine` into `GobalResolution`,
which is a much more sensible place for it, but that can wait
until I'm ready to start cleaning up `refinementUnit.ml`; at
this point I'm still focused on cleaning up and then
encapsulating the inner logic of `Annotation`.

NOTE: I'm pretty sure that this diff is actually a perf
win. Inline comments describe why, but I would argue that
the underlying problem was that the old code smashed together
the actual operations and a lot of `GlobalResolution` boilerplate
to set up the solver, which made it too hard to read and optimize.
Keeping the core logic separate from the boilerplate makes it easier
to see the perf implications of the logic.

Reviewed By: dkgi

Differential Revision: D32413759

fbshipit-source-id: bea7dca6f579e26875be47b7743ad354f186363a