RefinementUnit refactor [11] Fix meet, join
Summary:
The `RefinementUnit.t` type is a tree.

I'm going to argue that as a result, a reasonable implementation of
`meet` and `join` will be recursive. The code we had was not directly
recursive, it used a strange combination of two helper functions (one
of which was recursive) instead.

The renames from the previous commit made it clearer to me what was
happening, and so at this point it's clear that we can just use a
straightforward recursion, where:
- first we use a match against the bases to define both
  `base` and a short-circuiting flag `should_recurse`
- then, if `should_recurse` is true we use `fold2` to recursively
  meet / join the attributes
- we return the base and the attributes

Reviewed By: dkgi

Differential Revision: D32422079

fbshipit-source-id: 622c45ee8a730d71110d2d5698395c82c8762765