RefinementUnit refactor [9] Fix perf bug
Summary:
As discussed in the previous commit, the short-circuiting
code in `valid_{join,meet}` was actually buggy - we needed
to be looking at the attributes at the current layer of
the `RefinementUnit` tree, but we were instead constantly
looking at just the very top of the tree.

This commit fixes the bug, which requires that we pass
the entire `RefinementUnit.t` to the `valid_{join,meet}`
functions instead of just the bases.

Unit tests don't change because the bug was a perf bug,
not a logical bug: we were just failing to short-circuit
properly, and instead doing the full `fold2` when we
didn't need to.

NOTE: sadly this commit makes the name aliasing worse
again since `valid_{join,meet}` now take arguments
`left` and `right` which leads to 3 layers of aliasing
at one point inside the pattern matching on bases :(.
I'll clean that up in the next commit, in this commit
my focus is on a minimal fix for the actual bug rather than
continuously improving readability.

Reviewed By: dkgi

Differential Revision: D32419176

fbshipit-source-id: 23b3e3b00682fd23949efc5c3e549a77e5a83d3f