Add do-not-ignore-all-errors-in as cli argument
Summary:
Support this configuration field as a cli argument.

Note - this works as-is with our existing infrastructure to convert command line arguments into configurations and pass to the backend.

However, for command line targets we already pass in the global root as a filter root, so adding more to the allowlist is a no-op unless we remove the global root in the presence of this cli arg.

Reviewed By: grievejia

Differential Revision: D32440362

fbshipit-source-id: 46b7e50cba82c5ac74d98e1ebe1a2bf4ab89901a