In pyre infer, print uncaught exception tracebacks.
Summary:
In general, we have a lot of `pyre` code set up in a way that
hides exception details from users. The merits of this are
debateable for open-source projects where there's no real
reason to hide tracebacks, but it's a common practice in stable
code.

But `pyre infer` is much, much less stable than pyre, and it
hard-blocks `pysa` analysis in some CI setups.

Moreover, it's not always reasonable to rely entirely on local
debugging:
- firstly, infer-for-pysa can be pretty slow because it infers
  the *entire* codebase (as it should for pysa purposes) so reproducing
  an error on the python side locally can take a long, long time.
- secondly, reproducing erors insn't always reliable. Sometimes an
  error might only happen in a CI system. This is the case with
  a utf-8 decode error I'm investigating: running the exact same
  command locally did not produce the same error, so I have no way
  to debug. I *really* need to see where in the python stack the error
  is coming from, at the moment I can't tell if it's from parsing
  backend output or parsing python code on disk.

Reviewed By: arthaud

Differential Revision: D32468689

fbshipit-source-id: a125bce817146b9fe1a73fe891272d4d1ee3554f