Make inline decorator function take a callback to get decorator body.
Summary:
Problem: Want to switch to inlining at the preprocessing phase. In that design, we will get the decorator body from the source. So, change the signature for the inline decorator function.

# Overall design for the stack

Problem: Inlining a decorator `foo.bar.my_decorator` requires the body of the decorator. Right now, we use the global environment to look up the body. But this requires a subsequent incremental recheck and makes us take ~10 minutes more and uses extra memory even when no decorator is inlined.

Solution: We know by now that decorator inlining can be done as a source-transform without type environment knowledge. So, move it to the preprocessing stage.

Complication 1: Inlining in a file needs to know the sources of the decorators it uses. That means we need to come after those sources are preprocessed and fetch them on demand. We don't want to add a new phase just for inlining.

Approach: Preprocessing is done in two sub-phases - (a) those preprocessing steps that need to happen before wildcard import expansion (e.g., relative import expansion) (b) those that can come afterwards. We will inline between step (a) and (b), just like wildcard expansion, which also needs to fetch sources based on qualifiers.

Complication 2: Need to get the skipped decorators from the models file. But we do that using the type environment, which comes after the preprocessing phase.

Approach: Parse the skip-decorator model files separately for now. In the future, if we decouple model-parsing from the type environment, we can use that here.

Reviewed By: arthaud

Differential Revision: D32367802

fbshipit-source-id: bd3706c0e7ec6995095d0e63d0b3d484d87bed64