Move out computation of module reference.
Summary:
Problem: In issues, we wanted to show filename for an inlined decorator as its original file. So, we looked up the module using global resolution and stored a mapping from inlined name to module.

But once we move to the preprocessing phase, we won't have access to the global resolution.

In any case, the module is computable from the decorator. There is no real need to store it separately.

So, move the module fetching to the use site.

Reviewed By: arthaud

Differential Revision: D32367800

fbshipit-source-id: fc20d2f75b65e01ab1ae0230477d8038f57803b7