Remove decorators from infer [1/3] quick fix
Summary:
**Context**

We have a case where a utf-8 encoding bug on some Chinese
text inside a decorator is causing the `pyre infer` backend
to send a sequence of non-utf-8-decodable bytes to the python
side.

This is an actual bug we should address, but in the meantime
we need to get `pyre infer` working again ASAP so that we can infer
types for pysa runs.

**This diff**

`pyre infer` doesn't actually care about decorators at all, the
decorators pasted into the stub files are completely ignored by
`ApplyTypeAnnotationsVisitor`.

So, let's just stop dealing with them which will fix the utf-8
bug. As a quick fix, I'm just hardcoding the decorators to an
empty list so that we don't have to worry about updating the
client and such in order to start pushing a pysa release.

I'll follow up with a pair of diffs removing decorators from
the client and the backend.

Reviewed By: pradeep90

Differential Revision: D32475878

fbshipit-source-id: eeeb8e0f35d1fce0777bd55f80c84dfe18640486