Extract the taint for the implicit self before handling call targets in the forward analysis
Summary:
The way we handle method calls is pretty messy right now.
When the callee is unknown, we analyze the callee as a whole and propagate it.
When the callee is a method, we extract self and analyze it.
This diff refactors this handling, properly extracting the taint for a possible implicit `self` in a single place.

This will also make it possible to have callees with different `implicit_self` values, instead of ignoring them.

Reviewed By: dkgi

Differential Revision: D32405255

fbshipit-source-id: c8749da5bdf45ffb984f64abc54f6913aaa2bad9