Scope: capture / binding
Summary:
# This diff

Collect captures from the patterns & bodies of each case in a match statement.

Within the patterns, the captures directly occur in:

* MatchAs
* MatchMapping
* MatchStar

In the others it is via subpatterns.

The location for rest should be narrower in the case of MatchMapping's rest.

# Broader Context

Not necessary to understand this diff, but if interested, see D32027190 (https://github.com/facebook/pyre-check/commit/0d16c9b5206fefc2d00327df9078703eef93f76f)

Reviewed By: grievejia

Differential Revision: D32397852

fbshipit-source-id: d01dd48f018b8034e01c8cd826474f715e064dd3