Only include more specific of overlapping checked directory paths
Summary:
Specifying a `do_not_ignore_all_errors_in` directory path is supported via cli or as a field in the configuration, and should enable error reporting in that directory. Paths specified this way are augmented with an inferred allow list path, either the local root or the global root.

In most cases, explicit allow list paths point at a directory that is not included in the local root. However, the global root encompasses the entire repository and specifying any additional explicit allow list paths (ie., the command line use case) have no effect in actually narrowing the allow list at all.

To make it possible to specify a more specific allow list, while keeping the behavior of the cli flag and configuration field consistent, let's only include the inferred allow list path if it is not a prefix of an explicitly given path. This should allow users to provide a more specific explicit path and actually see a behavior change.

Reviewed By: dkgi

Differential Revision: D32449528

fbshipit-source-id: df40a12ec9e4912089017123ba613a4847f56567