Store decorators to skip before type-checking.
Summary:
Problem: We want to know which decorators to skip. The previous diff added code to parse models and find such decorators.

In this diff, we store the decorators in a global variable so that they are available when preprocessing.

Design choice: I'm putting it in a global variable because it would be pretty awkward to thread the decorators list to the preprocessing phase in the current API.

Reviewed By: arthaud

Differential Revision: D32466647

fbshipit-source-id: aefc2c53d7dea8981968806a8d02210f103a6620