Expose function to get decorator body.
Summary:
Need this in `query.ml`. That is the blocker to deleting the old decorator code.

Note that I'm slightly changing the caching in decorator body. The cached set of decorators for a module stores all decorators without skipping. That way, `query` can call with different `decorators_to_skip` without failing because a previous query skipped all decorators. Perf should not be affected, but will keep an eye out.

Reviewed By: dkgi

Differential Revision: D32473196

fbshipit-source-id: eb4a93cac01bdb1093bd42afcc060375b27a1f76