Delete old code for getting decorators to skip.
Summary:
The previous diff started using the new way of getting decorators to skip.

This means we no longer need to get part of the initialized models before inlining and then get the rest later. We can go back to the old way of getting the models at the same time.

Reviewed By: arthaud

Differential Revision: D32466650

fbshipit-source-id: f29a21f216467ac55a2aebdf80bb44a66eaa24df