Move annotation tests [1] join and meet
Summary:
As with the preceding commit, I'm moving tests of the
`Annotation.t` lattice functionality into `AnnotationTest`,
which makes the code shorter and clearer.

This also lets us strip the `RefinementUnit` tests to just
making sure the recursive operations are correctly applied,
because the various cases for what happens inside of
`Annotation.{join,meet}` is tested elsewhere. That makes the
tests easier to navigate and extend (as I will need to extend
them shortly for index refinement).

Reviewed By: grievejia

Differential Revision: D32497183

fbshipit-source-id: aa51c6fe29798bd523588a31fc87ada64cd9f04a