Clean up RefinementUnit create and test code
Summary:
In terms of the actual logic being in a reasonable state, I'm finally ready to begin working
on adding index-based refinement to `RefinementUnit`.

But the tests are still pretty bad - it's like some ugly emacs lisp exploded into
`refinementUnitTest.ml`. It's pretty hard to see the logic we're testing because the
boilerplate drowns it out

To clean it up, I do the following:
- Split the overloaded `RefinementUnit.create` function into a `create` and `empty`,
  where `create` takes a single required `Annotation.t` positional argument
- Make a `create_mutable` function that takes a `Type.t`. This makes quite a bit of
  the prod code simpler, but the main reason I did it is it makes the tests
  waaaayyyy easier to read, the prod code is just icing on the cake.
- Change `add_attribute_refinement` to take the refined `Annotation.t` as `annotation`
  rather than `base` becauese it's a little weird to call the thing `base` when we
  aren't actually updating `base`.
- Make a private helper function `add_mutable_attribute_refinement` so that all the
  tests that follow a `create |> add_some_attribute_refinements` are more compact;
  we didn't really need named args for this in the tests because it's super obvious
  what the `reference` component is.

NOTE: this is probably also a slight perf win for both CPU and memory, because now
`empty` is a static object and every call to `create` will reuse the same empty attribute
maps. That saves some allocations and RAM, it also might boost cache consistency
because whenever there are no attribute refinements (which is most of the time) we'll
be fetching the map from the same memory location.

Reviewed By: grievejia

Differential Revision: D32500437

fbshipit-source-id: 690f189d68338c23b0f01ad256379b5dbe00b0da