Don't use RefinementUnit to compare Annotations
Summary:
I'm not sure the historical context for the code, but we were
creating `RefinementUnit` values with only a base in order to
use `less_or_equal` in several places in `typeCheck.ml`.

This is weird, if we just want to compare `Annotation` values
we should use `Annotation.less_or_equal`.

This should save quite a bit of perf by skipping allocations
and map comparisons that were totally unnecessary. It also
makes the code quite a bit clearer.

While I was at it I renamed `annotation` to `type_` in one
spot where it was particularly confusing because there were
both `Type.t` and `Annotation.t` values floating around at
the same time.

Reviewed By: dkgi

Differential Revision: D32523579

fbshipit-source-id: eda676d9fa23d0a1962d2ee39db3fa0ce0a0f655