Improve the call graph representation to allow mixed calls of functions and methods
Summary:
This diff improves the call graph representation so that we can represent any resolution information.
Instead of returning a sum type with `RegularTargets`, `ConstructorTargets`, `HigherOrderTargets`, we use a product type, since any combination is possible.
We also introduce a boolean `unresolved` which represents whether at least one callee couldn't be resolved.

For now we keep the previous representation until we can fix the consumer part: the forward and backward taint analysis.

Reviewed By: dkgi

Differential Revision: D32516933

fbshipit-source-id: 4294ab49dc237222e9e704c70dacbfecd3e87f56