Make getattr(x, "field") behave same way as x.field
Summary:
# This diff

`getattr(x, "field")` in semantically equivalent to `x.field`; make type-checking logic aware of that. Before this diff, we would resolved the type as just `typing.Any`.

# Broader context

This is not required to understand the diff, but this will play a role in the stack starting at D32027190 (https://github.com/facebook/pyre-check/commit/0d16c9b5206fefc2d00327df9078703eef93f76f). In particular, it will come in handy for supporting `__match_args__`. Was able to shard this off as a more general standalone fix.

Reviewed By: pradeep90

Differential Revision: D32402339

fbshipit-source-id: c3d32b30aa8ab5a8af7961e10ba4081be7f19b98