Create Refinement.Store module
Summary:
The annotation_store logic is in pretty bad shape right now.

We have hacky lattice operations scattered through both
`typeCheck.ml` and `typeInferenceLocal.ml`, with bugs in them
because they've been copy-pasted and then gotten out of sync.

Moreover, for no real benefit we have the data structure
duplicated using trees so that it is shared-memory-safe inside
of localAnnotationMap.ml, which adds even more confusion.

We need a source of truth for the lattice, and the logical
place to group it is alongside `Refinement.Unit`, since it's
just a map of units.

This commit kicks off that migration by moving the data structure
itself, I'll migrate the lattices (which is where the payoff

Reviewed By: grievejia

Differential Revision: D32539565

fbshipit-source-id: a8fd6ca2213f0e8f2c01b4dc194407ddae559d72