Optimize MapLattice.less_or_equal to fold the right
Summary:
The existing `fold2`-based implementation of `less_or_equal` worked
and was easy to read, but as I looked at converting `typeCheck` and
`typeInferenceLocal` to using this code I realized that we were doing
a more efficient thing in some cases: because left having extra keys
is fine and `init` starts as `true`, we only actually need to fold
the right keys.

I made the doc comment more elaborate since the code is now less
self-explanatory than it was before (and actually the typeCheck code
from which I got this optimization did it wrong, see discussion and test
changes on the next commit - it's debatable whether this optimization
is even a good idea given that we've messed it up before, but I'm a little
afraid of a perf regression if I leave it out).

Reviewed By: grievejia

Differential Revision: D32543135

fbshipit-source-id: 9cc77253b3184dc36f6e0ced22da64ae61e2c39c