Add less_or_equal capability to Refinement.Store
Summary:
Move the implementations of less_or_equal, which were mostly duplicated
into inline portions of the less_or_equal on `TypeCheck.State` and
`TypeInferenceLocal.State`, into `Refinement.Store`.

There are two versions, `less_or_equal` which uses type ordering and
is relied on by the inference variant and `less_or_equal_monotone`
which uses equality checks on the `Unit` values (which is much, much, cheaper).

I'm going to see if anything breaks if I flip `typeInferenceLocal.ml` to
using the monotone variant - I don't think it will, and it should give us
a perf win as well as smaller code surface to maintain - but this diff
should remain a pure refactor.

NOTE: There's a bug in `TypeCheck.less_or_equal` here - left and right were
backward. It was really, really hard to see from the code (Shannon and I looked at
it for a while and eventually concluded that it was right, which is wrong!) but the bug
was pretty clear in the tests, where we were asserting that an empty `left` was
`less_or_equal` a nonempty `right`, which makes no sense.

**Why didn't we notice the bug**

The bug only affects performance, not correctness, which is why (almost) none of the
integration tests caught it: because the fixpoint is monotone, we were just going
until left and right were exactly equal. And to be honest it probably only even affected
performance in edge cases, since we'd have to be combining resolutions with different
sets of locals (this happens but is not super common in nicely-structured code).

Interestingly one integration test does actually catch it: an `Analysis failure [30]` in
`methodTest` disappears after this change. And the change seems legit - see line
comments, but I can't think of a reason why we should be complaining about the
annotation on `s` in the affected test, so I do think the bug actually caused some weird
edge cases.

Reviewed By: dkgi

Differential Revision: D32539564

fbshipit-source-id: 30de44fa6aed41e1c300a63b8597daf6cf606a9a