Move RefinementUnit to Refinement.Unit
Summary:
Our organization of refinement data is mostly okay up to the
`RefinementUnit` level, but `annotation_store` logic is defined
all over the place - much of it duplicated between `TypeCheck`
and `TypeInferenceLocal`, which is already a source of copypasta
code drift bugs and is just not acceptable.

Because there's quite a bit of overlap in functionality, I think
the best solution is to drop it into the same module as
`RefinementUnit`, so my intention is to put two submodules into
a `Refinement` module, so we'll have
- `Refinement.Unit` - refinements on all attributes of a given
  variable, without a distinction between temp and permanent
- `Refinement.Store` - refinements across all locally refined
  variables.

I may wind up with a private intermediate module `Refinement.Map`
but that won't be exposed if so.

The auto-renames for this commit were done as follows:
```
ack -l RefinementUnit | xargs xargs sed -i 's/RefinementUnit/Refinement.Unit/g'
```

Reviewed By: grievejia

Differential Revision: D32530435

fbshipit-source-id: d7b3ab95ef0ad767ada3932dd1bd2878b35c1464